static: move stuff in ui.js
